feat: 10/27/2023 update Packages that have updated: - aylurs-gtk-shell-git - grimblast-git - hyprland-git
Apologies for the delays in packaging. From today onwards packaging and updates will reschedule as normal. NixOS, while a fascinating concept, has poor documentation and would take far too much of my time to learn at this stage of my life. This repo will NOT be sunset, and will continue to be supported for much longer.